requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.9
django-cors-headers==3.10.0
django-filter==21.1
django-versatileimagefield==2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-flex-fields==0.9.5
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.2
Pillow==8.4.0
pycodestyle==2.8.0
PyJWT==2.3.0
pyparsing==2.4.7
python-decouple==3.5
python-http-client==3.3.3
python-magic==0.4.24
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
sendgrid==6.9.0
sendgrid-python==0.1.1
sqlparse==0.4.2
starkbank-ecdsa==1.1.1
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
